+++ CHANGELOG.md
@@ -4,8 +4,10 @@ All notable changes to the sdc11073 module will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.1.5] - 2021-11-16
+## [1.1.5] - 2021-11-18
 ### Fixed
+- use MetadataVersion in Wsdiscovery
+- fix setLocation for SdcDevice implementation - no Bye-message on location change
 - fixed comparison regarding float precision in NumericMetricValue.__eq__ and SampleArrayValue.__eq__
 
 ## [1.1.4] - 2021-10-15